fix: 리베이스 과정에서 오류해결
