+++ .gitignore
@@ -61,3 +61,4 @@ atlassian-ide-plugin.xml
 
 build-test.gradle
 /module-external-api/src/main/resources/firebase/q-feed-firebase-adminsdk-ig40w-d0bf1986b6.json
+/module-external-api/src/main/resources/firebase/q-feed-firebase-adminsdk-ig40w-2d9ccc78d3.json
