--- module-external-api/src/main/java/com/wsws/moduleexternalapi/fcm/dto/FCMRequestDto.java
@@ -1,7 +1,0 @@
-package com.wsws.moduleexternalapi.fcm.dto;
-
-public record FCMRequestDto(
-        String title,             // 알림 제목
-        String body,              // 알림 내용
-        String recipientFcmToken // 수신자 FCM 토큰
-) {}
