+++ module-external-api/src/main/java/com/wsws/moduleexternalapi/fcm/service/FcmService.java
@@ -4,11 +4,8 @@ import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.Message;
 import com.google.firebase.messaging.Notification;
-import com.wsws.moduledomain.user.User;
-import com.wsws.moduledomain.user.repo.UserRepository;
-import com.wsws.moduledomain.user.vo.UserId;
-import com.wsws.moduleexternalapi.fcm.dto.FCMRequestDto;
-//import com.wsws.moduleinfra.FcmRedis;
+import com.wsws.moduleexternalapi.fcm.dto.fcmRequestDto;
+import com.wsws.moduleinfra.FcmRedis;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -20,31 +17,22 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class FcmService {
 
-//      private final FcmRedis fcmRedis;
-      private final UserRepository userRepository;
+      private final FcmRedis fcmRedis;
 
-//    @Async("taskExecutor")
-//    public void FcmSend(String recipient, FCMRequestDto fcmRequestDto) {
-//        User user = getUserById(recipient); // 사용자 조회
-//        UserId userId = user.getId(); // UserId 가져오기
-//
-//        String fcmRedisKey = getFcmRedisKey(recipient); // Redis 키 생성.
-//        String fcmToken = fcmRedis.getFcmToken(fcmRedisKey); // Redis에서 FCM 토큰 조회.
-//
-//        if (fcmToken != null && !fcmToken.isEmpty()) { // 토큰이 존재하면 메시지 생성 후 전송.
-//            Message message = makeMessage(fcmRequestDto, fcmToken);
-//            sendMessage(message);
-//        }
-//    }
+    @Async("taskExecutor")
+    public void fcmSend(String recipient, fcmRequestDto fcmRequestDto) {
 
-    private User getUserById(String recipient) {
-        return userRepository.findById(UserId.of(recipient))
-                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 사용자 ID: " + recipient));
+        String fcmRedisKey = getFcmRedisKey(recipient); // Redis 키 생성.
+        String fcmToken = fcmRedis.getFcmToken(fcmRedisKey); // Redis에서 FCM 토큰 조회.
+
+        if (fcmToken != null && !fcmToken.isEmpty()) { // 토큰이 존재하면 메시지 생성 후 전송.
+            Message message = makeMessage(fcmRequestDto, fcmToken);
+            sendMessage(message);
+        }
     }
 
-
     // 메시지 생성
-    public Message makeMessage(FCMRequestDto fcmRequestDto, String token){ // 나중에 토큰도 추가
+    public Message makeMessage(fcmRequestDto fcmRequestDto, String token){ // 나중에 토큰도 추가
         Notification.Builder notificationBuilder =
                 Notification.builder()
                         .setTitle(fcmRequestDto.title())
@@ -65,6 +53,10 @@ public class FcmService {
         }
     }
 
+    public fcmRequestDto makeFcmDTO(String title, String body) {
+        return new fcmRequestDto(title, body);
+    }
+
     private String getFcmRedisKey(String userId) {
         return "FCM_TOKEN_" + userId; // Redis 키 형식 정의
     }
@@ -75,7 +67,7 @@ public class FcmService {
     }
 
     // 팔로우 알림 본문 생성
-    public String makeFollowBody(String sender, String reportMember, String type) {
+    public String makeFollowBody(String sender, String type) {
         return sender
                 + " 님이 회원님을" + type + " 했습니다." ;
     }
